pyproject.toml
[tool.poetry]
name = "perfect_chord"
version = "0.1.0"
description = "Web application for easy reading of chords when playing"
authors = ["Eugene Denkevich <e.denkevich@data-light.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
pydantic = "2.7.3"
fastapi = "0.111.0"
email-validator = "2.1.1"
passlib = "1.7.4"
uvicorn = "0.30.5"
gunicorn = "22.0.0"
loguru = "0.7.2"
pydantic-settings = "2.4.0"
sqlalchemy = "2.0.31"
alembic = "1.13.2"
asyncpg = "0.29.0"
pyjwt = "2.9.0"
python-dateutil = "2.9.0"
types-python-dateutil = "2.8.19.20240106"
cryptography = "43.0.0"

[tool.poetry.group.linters.dependencies]
pre-commit = "3.7.1"
ruff = "0.4.8"
mypy = "1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
exclude = [
    "backend/infrastructure/postgresql/migrations/versions/",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = [
    "F",
    "C",
    "E",
    "T10",
    "T20",
    "Q",
    "RET",
    "I"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"example/" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = [
    "backend",
]
